Update main.py
The bot now downloads files from their original format. It previously only downloaded images in a .png format, now allows .mp4 and .gif. formats. While uuid provided unique filenames for each file the bot downloads, it did not keep the files in chronological order. To fix this I changed the parameter for 'ctx.channel.history(limit = 200)' to 'oldest_first = True', this ensures that the bot downloads each image from first message to the last message. The naming convention for each file is the text channel name + the image index. The index is stored in a counter that increases after each loop. Use 'limit' if you want the bot to download in the opposite order.